+++ catalog/app/containers/Search/Search.tsx
@@ -159,7 +159,6 @@ function NumberFilterWidget({
   extents,
   onChange,
 }: FilterWidgetProps<SearchUIModel.Predicates['Number']>) {
-  // XXX: query extents
   const handleChange = React.useCallback(
     (value: { min: number | null; max: number | null }) => {
       onChange({ ...state, gte: value.min, lte: value.max })
@@ -192,7 +191,6 @@ function DatetimeFilterWidget({
   extents,
   onChange,
 }: FilterWidgetProps<SearchUIModel.Predicates['Datetime']>) {
-  // XXX: query extents
   const fixedExtents = React.useMemo(
     () => ({
       min: extents?.min ?? new Date(),
@@ -229,7 +227,6 @@ function KeywordEnumFilterWidget({
   extents,
   onChange,
 }: FilterWidgetProps<SearchUIModel.Predicates['KeywordEnum']>) {
-  // XXX: query extents
   const handleChange = React.useCallback(
     (value: string[]) => {
       onChange({ ...state, terms: value })
@@ -263,7 +260,7 @@ function KeywordWildcardFilterWidget({
   return (
     <FiltersUI.TextField
       onChange={handleChange}
-      placeholder="Match against (wildcards supproted)"
+      placeholder="Match against (wildcards supported)"
       value={state.wildcard}
     />
   )
@@ -336,11 +333,7 @@ interface PackagesFilterActivatorProps {
 }
 
 function PackagesFilterActivator({ field }: PackagesFilterActivatorProps) {
-  const model = SearchUIModel.use()
-  invariant(
-    model.state.resultType === SearchUIModel.ResultType.QuiltPackage,
-    'Filter type mismatch',
-  )
+  const model = SearchUIModel.use(SearchUIModel.ResultType.QuiltPackage)
   const { activatePackagesFilter } = model.actions
   const activate = React.useCallback(() => {
     activatePackagesFilter(field)
@@ -354,11 +347,7 @@ interface PackagesFilterProps {
 }
 
 function PackagesFilter({ className, field }: PackagesFilterProps) {
-  const model = SearchUIModel.use()
-  invariant(
-    model.state.resultType === SearchUIModel.ResultType.QuiltPackage,
-    'Filter type mismatch',
-  )
+  const model = SearchUIModel.use(SearchUIModel.ResultType.QuiltPackage)
   const predicateState = model.state.filter.predicates[field]
   invariant(predicateState, 'Filter not active')
 
@@ -425,11 +414,7 @@ function PackagesMetaFilterActivator({
   label,
   disabled,
 }: PackagesMetaFilterActivatorProps) {
-  const model = SearchUIModel.use()
-  invariant(
-    model.state.resultType === SearchUIModel.ResultType.QuiltPackage,
-    'Filter type mismatch',
-  )
+  const model = SearchUIModel.use(SearchUIModel.ResultType.QuiltPackage)
   const { activatePackagesMetaFilter } = model.actions
   const type = SearchUIModel.PackageUserMetaFacetMap[typename]
   const activate = React.useCallback(() => {
@@ -444,11 +429,7 @@ interface PackageMetaFilterProps {
 }
 
 function PackagesMetaFilter({ className, path }: PackageMetaFilterProps) {
-  const model = SearchUIModel.use()
-  invariant(
-    model.state.resultType === SearchUIModel.ResultType.QuiltPackage,
-    'Filter type mismatch',
-  )
+  const model = SearchUIModel.use(SearchUIModel.ResultType.QuiltPackage)
 
   const predicateState = model.state.userMetaFilters.filters.get(path)
   invariant(predicateState, 'Filter not active')
@@ -468,7 +449,7 @@ function PackagesMetaFilter({ className, path }: PackageMetaFilterProps) {
 
   const title = React.useMemo(() => JSONPointer.parse(path).join(' / '), [path])
 
-  const extents = SearchUIModel.usePackageUserMetaFacetExtents(path)
+  const { fetching, extents } = SearchUIModel.usePackageUserMetaFacetExtents(path)
 
   return (
     <FiltersUI.Container
@@ -477,7 +458,14 @@ function PackagesMetaFilter({ className, path }: PackageMetaFilterProps) {
       onDeactivate={deactivate}
       title={title}
     >
-      <FilterWidget state={predicateState} extents={extents} onChange={change} />
+      {fetching ? (
+        <>
+          <Skeleton height={32} />
+          <Skeleton height={32} mt={1} />
+        </>
+      ) : (
+        <FilterWidget state={predicateState} extents={extents} onChange={change} />
+      )}
     </FiltersUI.Container>
   )
 }
@@ -508,7 +496,7 @@ const useFilterGroupStyles = M.makeStyles((t) => ({
 interface FilterGroupProps {
   disabled?: boolean
   path?: string
-  items: ReturnType<typeof SearchUIModel.groupFacets>[number]['children']
+  items: SearchUIModel.FacetTree['children']
 }
 
 function FilterGroup({ disabled, path, items }: FilterGroupProps) {
@@ -569,14 +557,13 @@ function FilterGroup({ disabled, path, items }: FilterGroupProps) {
   )
 }
 
-const FACETS_THRESHOLD = 5
-
 const useAvailablePackagesMetaFiltersStyles = M.makeStyles((t) => ({
   list: {
     background: t.palette.background.default,
   },
   help: {
     ...t.typography.caption,
+    marginBottom: t.spacing(1),
     marginTop: t.spacing(1),
   },
   input: {
@@ -590,72 +577,70 @@ const useAvailablePackagesMetaFiltersStyles = M.makeStyles((t) => ({
 
 interface AvailablePackagesMetaFiltersProps {
   className?: string
+  filtering: SearchUIModel.FacetsFilteringStateInstance
+  facets: {
+    available: readonly SearchUIModel.PackageUserMetaFacet[]
+    visible: SearchUIModel.FacetTree
+    hidden: SearchUIModel.FacetTree
+  }
   fetching: boolean
-  filters: SearchUIModel.PackageUserMetaFacet[]
 }
 
 function AvailablePackagesMetaFilters({
   className,
+  filtering,
+  facets,
   fetching,
-  filters,
 }: AvailablePackagesMetaFiltersProps) {
   const classes = useAvailablePackagesMetaFiltersStyles()
-  const [query, setQuery] = React.useState('')
-  const filtered = React.useMemo(
-    () =>
-      filters.length > FACETS_THRESHOLD
-        ? filters.filter((f) =>
-            (f.path + SearchUIModel.PackageUserMetaFacetMap[f.__typename])
-              .toLowerCase()
-              .includes(query.toLowerCase()),
-          )
-        : filters,
-    [filters, query],
-  )
-
-  const filteredNumber = filtered.length
-  const hiddenNumber = filters.length - filteredNumber
 
   const [expanded, setExpanded] = React.useState(false)
   const toggleExpanded = React.useCallback(() => setExpanded((x) => !x), [])
 
-  const [head, tail] = React.useMemo(
-    () => SearchUIModel.groupFacets(filtered, FACETS_THRESHOLD),
-    [filtered],
-  )
-
   return (
     <div className={className}>
-      {filters.length > FACETS_THRESHOLD && (
-        <FiltersUI.TinyTextField
-          placeholder="Find metadata filter"
-          fullWidth
-          value={query}
-          onChange={setQuery}
-          className={classes.input}
-          disabled={fetching}
-        />
-      )}
+      {SearchUIModel.FacetsFilteringState.match({
+        Enabled: ({ value, set }) => (
+          <FiltersUI.TinyTextField
+            placeholder="Find metadata"
+            fullWidth
+            value={value}
+            onChange={set}
+            className={classes.input}
+          />
+        ),
+        Disabled: () => null,
+      })(filtering)}
+      {SearchUIModel.FacetsFilteringState.match({
+        Enabled: ({ isFiltered, serverSide }) => {
+          if (serverSide && !isFiltered) {
+            return (
+              <p className={classes.help}>
+                Some metadata not displayed.
+                <br />
+                Enter search query to see more.
+              </p>
+            )
+          }
+          if (isFiltered && !facets.available.length) {
+            return <p className={classes.help}>No metadata found matching your query</p>
+          }
+        },
+        Disabled: () => null,
+      })(filtering)}
       <M.List dense disablePadding className={classes.list}>
-        <FilterGroup disabled={fetching} items={head.children} />
+        <FilterGroup disabled={fetching} items={facets.visible.children} />
         <M.Collapse in={expanded}>
-          <FilterGroup disabled={fetching} items={tail.children} />
+          <FilterGroup disabled={fetching} items={facets.hidden.children} />
         </M.Collapse>
       </M.List>
-      {!!tail.children.size && (
+      {!!facets.hidden.children.size && (
         <MoreButton
           className={classes.more}
           disabled={fetching}
           onClick={toggleExpanded}
           reverse={expanded}
         />
-      )}
-      {!!hiddenNumber && (
-        <p className={classes.help}>
-          {filteredNumber
-            ? `There are ${hiddenNumber} more filters available. Loosen search query to see more.`
-            : `${hiddenNumber} filters are hidden. Clear search query to see them.`}
-        </p>
       )}
     </div>
   )
@@ -679,35 +664,34 @@ interface PackagesMetaFiltersProps {
 function PackagesMetaFilters({ className }: PackagesMetaFiltersProps) {
   const classes = usePackagesMetaFiltersStyles()
 
-  const { activatedPaths, available, fetching } = SearchUIModel.usePackagesMetaFilters()
+  const activated = SearchUIModel.use(SearchUIModel.ResultType.QuiltPackage).state
+    .userMetaFilters.filters
 
-  const noFilters = !(available.length + activatedPaths.length)
+  const activatedPaths = React.useMemo(() => Array.from(activated.keys()), [activated])
 
   return (
     <div className={className}>
-      <div className={classes.title}>
-        Package-level metadata
-        {fetching && <M.CircularProgress className={classes.spinner} size={12} />}
-      </div>
+      <div className={classes.title}>Package-level metadata</div>
       {activatedPaths.map((path) => (
         <FilterSection key={path}>
           <PackagesMetaFilter path={path} />
         </FilterSection>
       ))}
-      {!!available.length && (
-        <AvailablePackagesMetaFilters filters={available} fetching={fetching} />
-      )}
-      {noFilters && (
-        // XXX: nicer display
-        <M.Typography>No metadata found</M.Typography>
-      )}
+      <SearchUIModel.AvailablePackagesMetaFilters>
+        {SearchUIModel.AvailableFiltersState.match({
+          Loading: () => <M.Typography>Analyzing metadata&hellip;</M.Typography>,
+          Empty: () =>
+            activatedPaths.length ? null : <M.Typography>No metadata found</M.Typography>,
+          Ready: (ready) => <AvailablePackagesMetaFilters {...ready} />,
+        })}
+      </SearchUIModel.AvailablePackagesMetaFilters>
     </div>
   )
 }
 
-const packagesFiltersPrimary = ['workflow', 'modified'] as const
+const PACKAGES_FILTERS_PRIMARY = ['workflow', 'modified'] as const
 
-const packagesFiltersSecondary = ['size', 'name', 'hash', 'entries', 'comment'] as const
+const PACKAGES_FILTERS_SECONDARY = ['size', 'name', 'hash', 'entries', 'comment'] as const
 
 const usePackageFiltersStyles = M.makeStyles((t) => ({
   metadata: {
@@ -728,18 +712,13 @@ interface PackageFiltersProps {
 }
 
 function PackageFilters({ className }: PackageFiltersProps) {
-  const model = SearchUIModel.use()
+  const model = SearchUIModel.use(SearchUIModel.ResultType.QuiltPackage)
   const classes = usePackageFiltersStyles()
-
-  invariant(
-    model.state.resultType === SearchUIModel.ResultType.QuiltPackage,
-    'wrong result type',
-  )
 
   const { order: activeFilters, predicates } = model.state.filter
 
-  const availableFilters = packagesFiltersPrimary.filter((f) => !predicates[f])
-  const moreFilters = packagesFiltersSecondary.filter((f) => !predicates[f])
+  const availableFilters = PACKAGES_FILTERS_PRIMARY.filter((f) => !predicates[f])
+  const moreFilters = PACKAGES_FILTERS_SECONDARY.filter((f) => !predicates[f])
 
   const [expanded, setExpanded] = React.useState(false)
   const toggleExpanded = React.useCallback(() => setExpanded((x) => !x), [])
@@ -784,7 +763,7 @@ function PackageFilters({ className }: PackageFiltersProps) {
   )
 }
 
-const objectFilterLabels = {
+const OBJECT_FILTER_LABELS = {
   modified: 'Last modified',
   size: 'Object size',
   ext: 'Extension',
@@ -798,16 +777,12 @@ interface ObjectsFilterActivatorProps {
 }
 
 function ObjectsFilterActivator({ field }: ObjectsFilterActivatorProps) {
-  const model = SearchUIModel.use()
-  invariant(
-    model.state.resultType === SearchUIModel.ResultType.S3Object,
-    'Filter type mismatch',
-  )
+  const model = SearchUIModel.use(SearchUIModel.ResultType.S3Object)
   const { activateObjectsFilter } = model.actions
   const activate = React.useCallback(() => {
     activateObjectsFilter(field)
   }, [activateObjectsFilter, field])
-  return <FiltersUI.Activator title={objectFilterLabels[field]} onClick={activate} />
+  return <FiltersUI.Activator title={OBJECT_FILTER_LABELS[field]} onClick={activate} />
 }
 
 interface ObjectsFilterProps {
@@ -816,11 +791,7 @@ interface ObjectsFilterProps {
 }
 
 function ObjectsFilter({ className, field }: ObjectsFilterProps) {
-  const model = SearchUIModel.use()
-  invariant(
-    model.state.resultType === SearchUIModel.ResultType.S3Object,
-    'Filter type mismatch',
-  )
+  const model = SearchUIModel.use(SearchUIModel.ResultType.S3Object)
   const predicateState = model.state.filter.predicates[field]
   invariant(predicateState, 'Filter not active')
 
@@ -862,16 +833,16 @@ function ObjectsFilter({ className, field }: ObjectsFilterProps) {
       className={className}
       defaultExpanded
       onDeactivate={deactivate}
-      title={objectFilterLabels[field]}
+      title={OBJECT_FILTER_LABELS[field]}
     >
       <FilterWidget state={predicateState} extents={extents} onChange={change} />
     </FiltersUI.Container>
   )
 }
 
-const objectsFiltersPrimary = ['modified', 'ext'] as const
+const OBJECTS_FILTERS_PRIMARY = ['modified', 'ext'] as const
 
-const objectsFiltersSecondary = ['size', 'key', 'content', 'deleted'] as const
+const OBJECTS_FILTERS_SECONDARY = ['size', 'key', 'content', 'deleted'] as const
 
 const useObjectFiltersStyles = M.makeStyles((t) => ({
   more: {
@@ -889,18 +860,13 @@ interface ObjectFiltersProps {
 }
 
 function ObjectFilters({ className }: ObjectFiltersProps) {
-  const model = SearchUIModel.use()
+  const model = SearchUIModel.use(SearchUIModel.ResultType.S3Object)
   const classes = useObjectFiltersStyles()
-
-  invariant(
-    model.state.resultType === SearchUIModel.ResultType.S3Object,
-    'wrong result type',
-  )
 
   const { order: activeFilters, predicates } = model.state.filter
 
-  const availableFilters = objectsFiltersPrimary.filter((f) => !predicates[f])
-  const moreFilters = objectsFiltersSecondary.filter((f) => !predicates[f])
+  const availableFilters = OBJECTS_FILTERS_PRIMARY.filter((f) => !predicates[f])
+  const moreFilters = OBJECTS_FILTERS_SECONDARY.filter((f) => !predicates[f])
 
   const [expanded, setExpanded] = React.useState(false)
   const toggleExpanded = React.useCallback(() => setExpanded((x) => !x), [])
@@ -1089,23 +1055,6 @@ function ResultsPage({ className, hits, cursor, resultType }: ResultsPageProps) 
   )
 }
 
-type NextPageQueryResult =
-  | ReturnType<typeof SearchUIModel.useNextPagePackagesQuery>
-  | ReturnType<typeof SearchUIModel.useNextPageObjectsQuery>
-
-interface NextPageQueryProps {
-  after: string
-  children: (result: NextPageQueryResult) => React.ReactElement
-}
-
-function NextPageObjectsQuery({ after, children }: NextPageQueryProps) {
-  return children(SearchUIModel.useNextPageObjectsQuery(after))
-}
-
-function NextPagePackagesQuery({ after, children }: NextPageQueryProps) {
-  return children(SearchUIModel.useNextPagePackagesQuery(after))
-}
-
 interface NextPageProps {
   after: string
   resultType: SearchUIModel.ResultType
@@ -1115,8 +1064,8 @@ interface NextPageProps {
 function NextPage({ after, className, resultType }: NextPageProps) {
   const NextPageQuery =
     resultType === SearchUIModel.ResultType.S3Object
-      ? NextPageObjectsQuery
-      : NextPagePackagesQuery
+      ? SearchUIModel.NextPageObjectsQuery
+      : SearchUIModel.NextPagePackagesQuery
   return (
     <NextPageQuery after={after}>
       {(r) => {